Add editor project

+++ src/raytracer/textures/texture.h
@@ -9,7 +9,8 @@ namespace raytracer
 	{
 	public:
 		virtual ~Texture() = default;
-		[[nodiscard]] virtual color value(double u, double v, const point3& p) const = 0;
+
+		[[nodiscard]] virtual color value(float u, float v, const point3& p) const = 0;
 	};
 
 } // raytracer
